Polished the libao output a bit while testing it on my linux laptop: 	- parsing -a driver:file for file devices (well, this could be extended to a general scheme to choose between different driver subsystems via -a system:file or system:device:file ...) 	- corrected the error handling: -1 is expected on error by mpg123.c 	- giving help message with list of available drivers when invalid one is specified
I verified with one testfile that -a wav:file does the same as -w file. I think that this is good;-)
Also, I'd like to mention that it's nice to measure audio subsystem overhead with comparing -t and -a null.
With the libao in place, I'm wondering if we still want our own alsa output... let's see...

svn path=/trunk/; revision=200