Bye, bye, libtldl.
Really, get it installed in your system or forget about it. It is mostly harmful to have the libltdl as a copy in mpg123.
We seem to get rpms with libltdl in it where there is a libtool rpm for that... needless nuisance. Get your libltdl and then mpg123 -- it's Free!


svn path=/trunk/; revision=1980