Importing a C version of getbits. That fixes layer1 decoding.
I am not sure what exactly went wrong with the macro, but the C code makes mpg123 a compliant layer 1 decoder again.
This should be no performance regression, at least it is none on my system. The C code is still in the header, ready to be inlined for the same result.
Maybe I will move all getbits macros to static C functions...


svn path=/trunk/; revision=2024