More verbosity on HTTP action (print all server responses with -vvv) and bail out early when relocation headers don't provide a new location.
This is related to issue 2904736: The server in that case does not give

Location: http://some/place

but

location: http://some/place

All other headers are uppercase as it should be, so, rather than making mpg123 fuzzy match the broken header, I decided to make it quickly error out with a better assessment of the situation.
The server should be fixed to put out correct headers.


svn path=/trunk/; revision=2450