Add mpg123_replace_reader to the list of largefile-sensitive functions. Also, add some commentry... the whole thing is still a mess.
I have the situation that libmpg123 can be binary-incompatible in a new version because large file support is enabled. Now, additionally, there is an incompatibility when upgrading from a largefile build to the new one with mpg123_replace_reader_64().
That one is really nasty, but it's been an API bug in the current mindset. What should one do?
I am contemplating making a new library interface that drops usage of off_t. No backwards compatibility... though, this is a pity, since this is just a custom problem of Solaris/Linux in 32 bit mode.
This SUCKS!


svn path=/trunk/; revision=2611