Adding tunable feed buffer pool, to avoid malloc/free all the time. See MPG123_FEEDPOOL and MPG123_FEEDBUFFER.
Note that this doesn't seem to bring direct benefit on GNU/Linux or Solaris (the latter perhaps in a multithreaded environment), but can be calming for people to have 18 mallocs instead of 360 for decoding an mpeg file. The other aspect is that you now actually can tune the minimal size of the feeder buffers, which can have some effect (if they are too small, at least). Also, this does only have relevance to the feeder API and the internal buffering of non-seekable streams, as that uses the same buffering code.


svn path=/trunk/; revision=2907