Do not delete tag info right away, just after potentially dangerous seeks.
Though, I do have to rethink that memory leaking protection. In practice, ID3v2 tags are only at the beginning and are parsed once, even seek to 0 seeks to the proper byte offset _after_ the tag. Trouble would only be tags in the middle of the stream ... which is not common, but possible.

svn path=/trunk/; revision=3325