libmpg123: count those bits and check before read (most of the time)
This is a big update touching the core of the decoding right before
actual synth. The side info and scale factors along with huffman
work are now littered with checks for the number of bits they
actually need and, if precomputation is not feasible, the number of
bits they would have read too much without the safety checks of
getbits.h

We get a lot more early error messages and bail outs now from the
do_layer() routines. I hope this settles possible frame data
bounds violations (read-only!) for good. The problem is that in
case of the huffman decoding, the checks should not be too strict,
as a bit of over-read is expected and those bits are stuffed back
at the end. If that over-read happens on the last huffman run and
hits the actual frame data boundary, simply some zeros are stuffed
back for no harm to follow.


svn path=/trunk/; revision=4370